test: Move GivenGrid_And_Alignment to runtime tests
